+++ src/imio/smartweb/common/upgrades.py
@@ -15,3 +15,9 @@ def reload_registry(context):
 def reload_actions(context):
     portal_setup = api.portal.get_tool("portal_setup")
     portal_setup.runImportStepFromProfile(PROFILEID, "actions")
+
+
+def reindex_searchable_text(context):
+    portal_catalog = api.portal.get_tool("portal_catalog")
+    portal_catalog.manage_reindexIndex(ids=["SearchableText"])
+    logger.info("Reindexed SearchableText index")